requirements.txt
absl-py==2.1.0
chex==0.1.88
dsrl==0.1.0
flax==0.8.5
gym==0.26.2
gymnasium==0.28.1
h5py==3.12.1
jaxlib==0.4.30
matplotlib==3.9.4

ml_collections==0.1.1
ml_dtypes==0.5.0
mujoco==2.3.0
numba==0.60.0
numpy==1.26.4
opt_einsum==3.4.0
optax==0.2.4
orbax-checkpoint==0.6.4
safety-gymnasium==0.4.0
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
tensorboardX==2.6.2.2
tensorflow-probability==0.23.0
tensorstore==0.1.69
tqdm==4.67.1
typing_extensions==4.12.2
wandb==0.19.1
